# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/extract:
    post:
      operationId: extractGeminiDocument
      tags: [gemini]
      summary: Extract allowed values from an uploaded document with Gemini
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiDocumentExtractionInput"
      responses:
        "200":
          description: Document extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiDocumentExtractionResult"
        "400":
          description: Invalid extraction request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Document exceeds the 8 MB inline-data limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gemini/repair-summary:
    post:
      operationId: generateGeminiRepairSummary
      tags: [gemini]
      summary: Generate bounded advisory guidance for a failed system event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiRepairSummaryInput"
      responses:
        "200":
          description: Advisory troubleshooting guidance grounded in the event snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiRepairSummaryResult"
        "400":
          description: Invalid repair-summary request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Repair-summary rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI provider did not return usable guidance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        category:
          type: string
        recoverable:
          type: boolean
...
    GeminiDocumentExtractionInput:
      type: object
      additionalProperties: false
      required: [contentBase64, mimeType, steps, destination, allowedFields]
      properties:
        contentBase64:
          type: string
          minLength: 4
          maxLength: 11184812
        mimeType:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[A-Za-z0-9][A-Za-z0-9!#$&^_.+-]*/[A-Za-z0-9][A-Za-z0-9!#$&^_.+-]*$"
        steps:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 2000
        destination:
          type: string
          maxLength: 2000
        allowedFields:
          type: array
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 200
        profileId:
          type: string
          maxLength: 200
        profileVersion:
          type: number
          minimum: 1
          multipleOf: 1
        documentType:
          type: string
          maxLength: 200
        examples:
          type: string
          maxLength: 4000
    GeminiDocumentExtractionResult:
      type: object
      additionalProperties: false
      required: [text, source, values]
      properties:
        text:
          type: string
        source:
          type: string
        values:
          type: object
          additionalProperties:
            type: string
    GeminiRepairSummaryEvent:
      type: object
      additionalProperties: false
      required: [eventId, timestamp, category, action, status, detail, actorSource, metadata]
      properties:
        eventId:
          type: string
          minLength: 1
          maxLength: 200
        timestamp:
          type: string
          maxLength: 80
        category:
          type: string
          maxLength: 100
        action:
          type: string
          maxLength: 200
        status:
          type: string
          maxLength: 80
        detail:
          type: string
          maxLength: 2000
        actorSource:
          type: string
          maxLength: 200
        operationId:
          type: string
          maxLength: 200
        metadata:
          type: object
          additionalProperties:
            type: string
            maxLength: 300
          maxProperties: 12
    GeminiRepairSummaryInput:
      type: object
      additionalProperties: false
      required: [event]
      properties:
        event:
          $ref: "#/components/schemas/GeminiRepairSummaryEvent"
    GeminiRepairSummaryResult:
      type: object
      additionalProperties: false
      required: [recordedEvidence, likelyCauses, repairActions, verificationSteps, advisory]
      properties:
        recordedEvidence:
          type: array
          minItems: 1
          maxItems: 8
          items:
            type: string
            minLength: 1
            maxLength: 1000
        likelyCauses:
          type: array
          minItems: 1
          maxItems: 8
          items:
            type: string
            minLength: 1
            maxLength: 1000
        repairActions:
          type: array
          minItems: 1
          maxItems: 8
          items:
            type: string
            minLength: 1
            maxLength: 1000
        verificationSteps:
          type: array
          minItems: 1
          maxItems: 8
          items:
            type: string
            minLength: 1
            maxLength: 1000
        advisory:
          type: string
          minLength: 1
          maxLength: 500