# lib/api-zod/src/generated/types/geminiRepairSummaryResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiRepairSummaryResult {
  /**
     * @minItems 1
     * @maxItems 8
     * @items.minLength 1
     * @items.maxLength 1000
     */
  recordedEvidence: string[];
  /**
     * @minItems 1
     * @maxItems 8
     * @items.minLength 1
     * @items.maxLength 1000
     */
  likelyCauses: string[];
  /**
     * @minItems 1
     * @maxItems 8
     * @items.minLength 1
     * @items.maxLength 1000
     */
  repairActions: string[];
  /**
     * @minItems 1
     * @maxItems 8
     * @items.minLength 1
     * @items.maxLength 1000
     */
  verificationSteps: string[];
  /**
     * @minLength 1
     * @maxLength 500
     */
  advisory: string;
}